requirements.txt
Flask
gunicorn
google-cloud-vision
numpy
opencv-python-headless
scipy
